refactor mcp presentation icons
